Support integers as status codes
Took a bit of digging but dtolnay was able to figure out how to support
allowing both strings and integers for status codes. This means that we
can now parse OpenAPI specs that are written in a more lenient way, i.e.
without a bunch of quotes in YAML.

Since all the fixtures are real specs (AFAICT), I didn't want to change
them and only added unit tests.
