+++ Cargo.toml
@@ -10,6 +10,6 @@ repository = "https://github.com/glademiller/openapiv3"
 description = "This crate aims to provide data structures that represent the Open API v3 specification easily deserializable with serde."
 
 [dependencies]
-serde = {version = "1.0", features = ["derive"]}
-serde_json = "1.0"
-serde_yaml = "0.8"
+serde = {version = "1.0.93", features = ["derive"]}
+serde_json = "1.0.39"
+serde_yaml = "0.8.9"
